### 2017-06-03 #276 @whedon
[PRE REVIEW]: Moltres: finite element based simulation of molten salt reactors
**Submitting author:** @lindsayad (<a href="http://orcid.org/0000-0002-6988-2123">Alexander Lindsay</a>)
**Repository:** <a href="https://github.com/arfc/moltres" target ="_blank">https://github.com/arfc/moltres</a>
**Version:** v0.1.0
