### 2017-06-05 #276 @kyleniemeyer
@arfon sure, I can take this one.

@lindsayad to start, can you recommend a potential reviewer or two? My preference is to get one set of eyes familiar with the topic/domain, and another more general reviewer.
