### 2017-06-05 #276 @lindsayad
Ok I'm thinking initial suggestions...@friedmud and @scopatz
