### 2017-06-06 #276 @kyleniemeyer
Hi @friedmud & @scopatz, would either (or both!) of you be willing to review this submission to JOSS?

From the JOSS reviewer pool, @tacaswell are you able/willing to review this?
