### 2017-06-06 #276 @scopatz
Hi @kyleniemeyer @whedon @arfon - definitely willing, but what is the timeline / deadline for the review? I am traveling a lot this summer.
