+++ src/base/MoltresApp.C
@@ -40,6 +40,7 @@
 #include "GenericMoltresMaterial.h"
 #include "CammiFuel.h"
 #include "CammiModerator.h"
+#include "Nusselt.h"
 
 // Postprocessors
 #include "AverageFissionHeat.h"
@@ -139,6 +140,7 @@ MoltresApp::registerObjects(Factory & factory)
   registerMaterial(GenericMoltresMaterial);
   registerMaterial(CammiFuel);
   registerMaterial(CammiModerator);
+  registerMaterial(Nusselt);
   registerPostprocessor(IntegralOldVariablePostprocessor);
   registerPostprocessor(ElementL2Diff);
   registerPostprocessor(IntegralNewVariablePostprocessor);
