+++ src/materials/GenericMoltresMaterial.C
@@ -16,16 +16,18 @@ validParams<GenericMoltresMaterial>()
                                "The number of delayed neutron precursor groups.");
   params.addCoupledVar(
       "temperature", 937, "The temperature field for determining group constants.");
-  MooseEnum interp_type("bicubic_spline spline least_squares none");
+  MooseEnum interp_type("bicubic_spline spline least_squares monotone_cubic none");
   params.addRequiredParam<MooseEnum>(
       "interp_type", interp_type, "The type of interpolation to perform.");
   params.addParam<std::vector<Real>>("fuel_temp_points",
                                      "The fuel temperature interpolation points.");
   params.addParam<std::vector<Real>>("mod_temp_points",
                                      "The moderator temperature interpolation points.");
-  params.addParam<PostprocessorName>("other_temp", 0, "If doing bivariable interpolation, need to "
-                                                      "supply a postprocessor for the average "
-                                                      "temperature of the other material.");
+  params.addParam<PostprocessorName>("other_temp",
+                                     0,
+                                     "If doing bivariable interpolation, need to "
+                                     "supply a postprocessor for the average "
+                                     "temperature of the other material.");
   params.addParam<std::string>("material", "Must specify either *fuel* or *moderator*.");
   params.addParam<bool>(
       "sss2_input", true, "Whether serpent 2 was used to generate the input files.");
@@ -120,6 +122,9 @@ GenericMoltresMaterial::GenericMoltresMaterial(const InputParameters & parameter
   else if (_interp_type == "spline")
     splineConstruct(property_tables_root, xsec_names);
 
+  else if (_interp_type == "monotone_cubic")
+    monotoneCubicConstruct(property_tables_root, xsec_names);
+
   else if (_interp_type == "bicubic_spline")
     bicubicSplineConstruct(property_tables_root, xsec_names, parameters);
 
@@ -193,6 +198,44 @@ GenericMoltresMaterial::splineConstruct(std::string & property_tables_root,
       for (decltype(o) k = 0; k < o; ++k)
         _xsec_spline_interpolators[xsec_names[j]][k].setData(temperature,
                                                              xsec_map[xsec_names[j]][k]);
+    }
+    else
+      mooseError("Unable to open file " + file_name);
+  }
+}
+
+void
+GenericMoltresMaterial::monotoneCubicConstruct(std::string & property_tables_root,
+                                               std::vector<std::string> xsec_names)
+{
+  Real value;
+  for (decltype(xsec_names.size()) j = 0; j < xsec_names.size(); ++j)
+  {
+    std::vector<Real> temperature;
+    std::string file_name = property_tables_root + _file_map[xsec_names[j]] + ".txt";
+    const std::string & file_name_ref = file_name;
+    std::ifstream myfile(file_name_ref.c_str());
+
+    auto o = _vec_lengths[xsec_names[j]];
+
+    std::map<std::string, std::vector<std::vector<Real>>> xsec_map;
+    xsec_map[xsec_names[j]].resize(o);
[…]
