+++ src/materials/GenericMoltresMaterial.C
@@ -16,7 +16,7 @@ validParams<GenericMoltresMaterial>()
                                "The number of delayed neutron precursor groups.");
   params.addCoupledVar(
       "temperature", 937, "The temperature field for determining group constants.");
-  MooseEnum interp_type("bicubic_spline spline least_squares monotone_cubic none");
+  MooseEnum interp_type("bicubic_spline spline least_squares monotone_cubic linear none");
   params.addRequiredParam<MooseEnum>(
       "interp_type", interp_type, "The type of interpolation to perform.");
   params.addParam<std::vector<Real>>("fuel_temp_points",
@@ -128,6 +128,9 @@ GenericMoltresMaterial::GenericMoltresMaterial(const InputParameters & parameter
   else if (_interp_type == "bicubic_spline")
     bicubicSplineConstruct(property_tables_root, xsec_names, parameters);
 
+  else if (_interp_type == "linear")
+    linearConstruct(property_tables_root, xsec_names);
+
   else if (_interp_type == "none")
     dummyConstruct(property_tables_root, xsec_names);
 
@@ -236,6 +239,44 @@ GenericMoltresMaterial::monotoneCubicConstruct(std::string & property_tables_roo
       for (decltype(o) k = 0; k < o; ++k)
         _xsec_monotone_cubic_interpolators[xsec_names[j]][k].setData(temperature,
                                                                      xsec_map[xsec_names[j]][k]);
+    }
+    else
+      mooseError("Unable to open file " + file_name);
+  }
+}
+
+void
+GenericMoltresMaterial::linearConstruct(std::string & property_tables_root,
+                                        std::vector<std::string> xsec_names)
+{
+  Real value;
+  for (decltype(xsec_names.size()) j = 0; j < xsec_names.size(); ++j)
+  {
+    std::vector<Real> temperature;
+    std::string file_name = property_tables_root + _file_map[xsec_names[j]] + ".txt";
+    const std::string & file_name_ref = file_name;
+    std::ifstream myfile(file_name_ref.c_str());
+
+    auto o = _vec_lengths[xsec_names[j]];
+
+    std::map<std::string, std::vector<std::vector<Real>>> xsec_map;
+    xsec_map[xsec_names[j]].resize(o);
+    _xsec_linear_interpolators[xsec_names[j]].resize(o);
+    if (myfile.is_open())
+    {
+      while (myfile >> value)
+      {
+        temperature.push_back(value);
+        for (decltype(o) k = 0; k < o; ++k)
+        {
+          myfile >> value;
+          xsec_map[xsec_names[j]][k].push_back(value);
+        }
+      }
+      myfile.close();
+      for (decltype(o) k = 0; k < o; ++k)
+        _xsec_linear_interpolators[xsec_names[j]][k].setData(temperature,
+                                                             xsec_map[xsec_names[j]][k]);
     }
     else
       mooseError("Unable to open file " + file_name);
@@ -510,6 +551,57 @@ GenericMoltresMaterial::monotoneCubicComputeQpProperties()
 }
 
 void
+GenericMoltresMaterial::linearComputeQpProperties()
+{
+  for (decltype(_num_groups) i = 0; i < _num_groups; ++i)
+  {
[…]
