+++ problems/MooseGold/property_file_dir/dummy_GTRANSFXS.txt
@@ -1,0 +1,3 @@
+1 0
+51 0
+101 0
