+++ problems/MooseGold/property_file_dir/dummy_DIFFCOEF.txt
@@ -1,0 +1,3 @@
+1 0.5
+51 0.5
+101 0.5
