+++ problems/MooseGold/property_file_dir/newt_one_group_msre_mod_REMXS.txt
@@ -1,0 +1,5 @@
+922 5.01273e-05
+972 4.99734e-05
+1022 4.9882e-05
+1072 4.97466e-05
+1122 4.96124e-05
