+++ problems/MooseGold/property_file_dir/newt_one_group_msre_mod_CHI.txt
@@ -1,0 +1,5 @@
+922 0.00000E+00
+972 0.00000E+00
+1022 0.00000E+00
+1072 0.00000E+00
+1122 0.00000E+00
