### 2017-06-15 845f6f6
A fair amount of new input plus update squirrel.
