### 2017-06-16 #276 @scopatz
It would be helpful if this link were somewhere in the original post: http://joss.theoj.org/about#reviewer_guidelines
