### 2017-06-16 #276 @scopatz
Also, it would be helpful if @lindsayad could provide links to the PDF of the paper and the API documentation.  Also also, is there a standard form or format for the review? It just seems like we are "encouraged to open issues against the repository." So it is not clear what a review truly entails. A little more formalism for the reviewer would be very helpful to me.
