### 2017-06-16 #276 @kyleniemeyer
Hi @scopatz, the main review format is based on the items given at the top, with the checkboxes. You can check off items if you feel they are satisfied—otherwise bring up comments here or issues in the source repo if something is not satisfied.
