### 2017-06-16 #276 @kyleniemeyer
@scopatz the aforementioned review checklist will be in #298, so let's move the conversation over there.
