### 2017-06-20 05b747c
Create FissionHeatSourceTransientAux so that we can find maximum (or average) power density in reactor.
