### 2017-06-20 b99b3e9
Implement delta_sigma_a proportional controller. Also discovered that line_search=none can vastly speed up convergence to steady-state.
