+++ include/auxkernels/FissionHeatSourceTransientAux.h
@@ -20,6 +20,7 @@ protected:
   const MaterialProperty<std::vector<Real>> & _fissxs;
   const MaterialProperty<std::vector<Real>> & _fisse;
   unsigned int _num_groups;
+  Real _nt_scale;
   std::vector<const VariableValue *> _group_fluxes;
 };
 
