### 2017-06-21 #298 @scopatz
For reference, the link to the "paper", which is really just a summary, is [here](https://github.com/arfc/moltres/blob/devel/paper/paper.pdf). Doxygen pages are [here](https://arfc.github.io/moltres/)
