### 2017-06-21 #298 @scopatz
Hi @lindsayad, there doesn't seem to be a conda package for moose. Can you confirm and/or create one? That would be the easiest way for me to verify the installation procedures. Thanks!
