### 2017-06-22 df0d449
Simulate true MSRE single channel flow and add opcode notes.
