+++ notes/opcode.md
@@ -1,0 +1,78 @@
+cAbs
+cAcos
+cAcosh
+cArg
+cAsin
+cAsinh
+cAtan
+cAtan2
+cAtanh
+cCbrt
+cCeil
+cConj
+cCos
+cCosh
+cCot
+cCsc
+cExp
+cExp2
+cFloor
+cHypot
+cIf
+cImag
+cInt
+cLog
+cLog10
+cLog2
+cMax
+cMin
+cPolar
+cPow
+cReal
+cSec
+cSin
+cSinh
+cSqrt
+cTan
+cTanh
+cTrunc
+cImmed
+cJump
+cNeg
+cAdd
+cSub
+cMul
+cDiv
+cMod
+cEqual
+cNEqual
+cLess
+cLessOrEq
+cGreater
+cGreaterOrEq
+cNot
+cAnd
+cOr
+cNotNot
+cDeg
+cRad
+cFCall
+cPCall
+cPopNMov
+cLog2by
+cNop
+cSinCos
+cSinhCosh
+cAbsAnd
+cAbsOr
+cAbsNot
+cAbsNotNot
+cAbsIf
+cDup
+cFetch
+cInv
+cSqr
+cRDiv
+cRSub
+cRSqrt
+VarBegin
