### 2017-06-22 15285c6
Number opcodes.
