+++ notes/opcode.md
@@ -1,78 +1,78 @@
-cAbs
-cAcos
-cAcosh
-cArg
-cAsin
-cAsinh
-cAtan
-cAtan2
-cAtanh
-cCbrt
-cCeil
-cConj
-cCos
-cCosh
-cCot
-cCsc
-cExp
-cExp2
-cFloor
-cHypot
-cIf
-cImag
-cInt
-cLog
-cLog10
-cLog2
-cMax
-cMin
-cPolar
-cPow
-cReal
-cSec
-cSin
-cSinh
-cSqrt
-cTan
-cTanh
-cTrunc
-cImmed
-cJump
-cNeg
-cAdd
-cSub
-cMul
-cDiv
-cMod
-cEqual
-cNEqual
-cLess
-cLessOrEq
-cGreater
-cGreaterOrEq
-cNot
-cAnd
-cOr
-cNotNot
-cDeg
-cRad
-cFCall
-cPCall
-cPopNMov
-cLog2by
-cNop
-cSinCos
-cSinhCosh
-cAbsAnd
-cAbsOr
-cAbsNot
-cAbsNotNot
-cAbsIf
-cDup
-cFetch
-cInv
-cSqr
-cRDiv
-cRSub
-cRSqrt
-VarBegin
+0. cAbs
+1. cAcos
+2. cAcosh
+3. cArg
+4. cAsin
+5. cAsinh
+6. cAtan
+7. cAtan2
+8. cAtanh
+9. cCbrt
+10. cCeil
+11. cConj
+12. cCos
+13. cCosh
+14. cCot
+15. cCsc
+16. cExp
+17. cExp2
+18. cFloor
+19. cHypot
+20. cIf
+21. cImag
+22. cInt
+23. cLog
+24. cLog10
+25. cLog2
+26. cMax
+27. cMin
+28. cPolar
+29. cPow
+30. cReal
+31. cSec
+32. cSin
+33. cSinh
+34. cSqrt
+35. cTan
+36. cTanh
+37. cTrunc
+38. cImmed
+39. cJump
+40. cNeg
+41. cAdd
+42. cSub
+43. cMul
+44. cDiv
+45. cMod
+46. cEqual
+47. cNEqual
+48. cLess
+49. cLessOrEq
+50. cGreater
+51. cGreaterOrEq
+52. cNot
+53. cAnd
+54. cOr
+55. cNotNot
+56. cDeg
+57. cRad
+58. cFCall
+59. cPCall
+60. cPopNMov
+61. cLog2by
+62. cNop
+63. cSinCos
+64. cSinhCosh
+65. cAbsAnd
+66. cAbsOr
+67. cAbsNot
+68. cAbsNotNot
+69. cAbsIf
+70. cDup
+71. cFetch
+72. cInv
+73. cSqr
+74. cRDiv
+75. cRSub
+76. cRSqrt
+77. VarBegin
