### 2017-06-22 #276 @friedmud
I'm a bit late to the party (been traveling) but I would be willing to do a
review as well.  Please advise if my review is needed at this point.

On Wed, Jun 21, 2017 at 1:10 PM Anthony Scopatz <notifications@github.com>
wrote:

> Thanks @lindsayad <https://github.com/lindsayad>!
>
