### 2017-06-22 #298 @Kevin-Mattheus-Moerman
@kyleniemeyer over at https://github.com/openjournals/joss-reviews/issues/276#issuecomment-310480159 @friedmud volunteered to review as well. Can you use another reviewer?
