### 2017-06-22 #298 @lindsayad
I would certainly be appreciated if @friedmud could review. As the creator of MOOSE and a developer of his own MOOSE applications, I think he could give some very helpful feedback...
