### 2017-06-22 #298 @lindsayad
@scopatz I will look into/get started on a conda package
