### 2017-06-23 #298 @kyleniemeyer
Hi @friedmud, your additional review is absolutely welcome. I believe @scopatz is taking care of checking the boxes above, but it would be great if you also go through each item and ensure you agree that they are satisfied (otherwise mention it here).
