### 2017-06-26 2dff7fc
Fix GenericMoltresMaterial
