### 2017-06-26 #298 @lindsayad
@scopatz So Moltres sits on top of MOOSE which does not do releases nor does it itself have a conda package. I could include MOOSE as a submodule of Moltres. However, MOOSE relies on a lot of dependencies like MPI and PetSc and needs environment variables like MPI_HOME and PETSC_DIR. Current MOOSE installation relies on providing a redistributable package which contains dependencies like MPI and PetSc and then automatically loads module files that modify things like `C_INCLUDE_PATH`, `CPLUS_INCLUDE_PATH`, `MPI_HOME`, `PETSC_DIR`, etc., creating the proper environment for building libmesh, MOOSE, and MOOSE applications. I've spent some time this afternoon with `conda build` and I believe I could eventually re-create the proper environment, but I think it might be kind of a pain.
