### 2017-06-27 #298 @lindsayad
@scopatz If it's of interest, and maybe of use for build questions, you can see an example of our automated build and test process [here](https://www.moosebuild.org/job/97233/)
