### 2017-06-27 #298 @scopatz
Hi @lindsayad can you please provide a VM or a docker container or something with all of the dependencies.  I know that it is kind of a pain to get all of the dependencies. I don't do moose development, and so I am looking for a way to evaluate Moltres without having to drill down the whole stack myself. Anything you can provide will be most useful.
