### 2017-06-28 #298 @lindsayad
Ok, the image is about 5.5 GB, but you can get an image with all the MOOSE dependencies installed as well as an already built moltres if it's needed for any reason. You can get it by running
```
docker pull lindsayad/moltres
```
