### 2017-07-05 8fe1534
Add new function material that provides slightly different results but speeds up two-d axi simulation by a factor of four. Also rewrite test so it's faster.
