+++ src/materials/MsreFuelTwoGrpXSFunctionMaterial.C
@@ -1,0 +1,156 @@
+#include "MsreFuelTwoGrpXSFunctionMaterial.h"
+#include "MooseUtils.h"
+// #define PRINT(var) #var
+
+template <>
+InputParameters
+validParams<MsreFuelTwoGrpXSFunctionMaterial>()
+{
+  InputParameters params = validParams<GenericConstantMaterial>();
+  params.addCoupledVar(
+      "temperature", 937, "The temperature field for determining group constants.");
+  return params;
+}
+
+MsreFuelTwoGrpXSFunctionMaterial::MsreFuelTwoGrpXSFunctionMaterial(
+    const InputParameters & parameters)
+  : GenericConstantMaterial(parameters),
+    _T(coupledValue("temperature")),
+    // _rho(getMaterialProperty<Real>("rho")),
+    _remxs(declareProperty<std::vector<Real>>("remxs")),
+    _fissxs(declareProperty<std::vector<Real>>("fissxs")),
+    _nsf(declareProperty<std::vector<Real>>("nsf")),
+    _fisse(declareProperty<std::vector<Real>>("fisse")),
+    _diffcoef(declareProperty<std::vector<Real>>("diffcoef")),
+    _recipvel(declareProperty<std::vector<Real>>("recipvel")),
+    _chi(declareProperty<std::vector<Real>>("chi")),
+    _gtransfxs(declareProperty<std::vector<Real>>("gtransfxs")),
+    _beta_eff(declareProperty<std::vector<Real>>("beta_eff")),
+    _beta(declareProperty<Real>("beta")),
+    _decay_constant(declareProperty<std::vector<Real>>("decay_constant")),
+
+    _d_remxs_d_temp(declareProperty<std::vector<Real>>("d_remxs_d_temp")),
+    _d_fissxs_d_temp(declareProperty<std::vector<Real>>("d_fissxs_d_temp")),
+    _d_nsf_d_temp(declareProperty<std::vector<Real>>("d_nsf_d_temp")),
+    _d_fisse_d_temp(declareProperty<std::vector<Real>>("d_fisse_d_temp")),
+    _d_diffcoef_d_temp(declareProperty<std::vector<Real>>("d_diffcoef_d_temp")),
+    _d_recipvel_d_temp(declareProperty<std::vector<Real>>("d_recipvel_d_temp")),
+    _d_chi_d_temp(declareProperty<std::vector<Real>>("d_chi_d_temp")),
+    _d_gtransfxs_d_temp(declareProperty<std::vector<Real>>("d_gtransfxs_d_temp")),
+    _d_beta_eff_d_temp(declareProperty<std::vector<Real>>("d_beta_eff_d_temp")),
+    _d_beta_d_temp(declareProperty<Real>("d_beta_d_temp")),
+    _d_decay_constant_d_temp(declareProperty<std::vector<Real>>("d_decay_constant_d_temp"))
+{
+}
+
+void
+MsreFuelTwoGrpXSFunctionMaterial::computeQpProperties()
+{
[…]
