+++ problems/MooseGold/single_msre_channel_velocity_only/build.txt
@@ -1,0 +1,56 @@
+Info    : Running 'gmsh -3 -order 2 half-fuel-channel-msre-dimensions.geo' [Gmsh 2.16.0, 1 node, max. 1 thread]
+Info    : Started on Fri Jun 30 11:55:42 2017
+Info    : Reading 'half-fuel-channel-msre-dimensions.geo'...
+Info    : Done reading 'half-fuel-channel-msre-dimensions.geo'
+Info    : Finalized high order topology of periodic connections
+Info    : Meshing 1D...
+Info    : Meshing curve 1 (Line)
+Info    : Meshing curve 2 (Line)
+Info    : Meshing curve 3 (Line)
+Info    : Meshing curve 4 (Line)
+Info    : Meshing curve 5 (Line)
+Info    : Meshing curve 6 (Line)
+Info    : Meshing curve 7 (Line)
+Info    : Meshing curve 8 (Line)
+Info    : Meshing curve 9 (Line)
+Info    : Meshing curve 10 (Line)
+Info    : Meshing curve 11 (Line)
+Info    : Meshing curve 12 (Line)
+Info    : Done meshing 1D (0.028 s)
+Info    : Meshing 2D...
+Info    : Meshing surface 14 (transfinite)
+Info    : Meshing surface 16 (transfinite)
+Info    : Meshing surface 18 (transfinite)
+Info    : Meshing surface 20 (transfinite)
+Info    : Meshing surface 22 (transfinite)
+Info    : Meshing surface 24 (transfinite)
+Info    : Done meshing 2D (0.000466108 s)
+Info    : Meshing 3D...
+Info    : Meshing volume 26 (transfinite)
+Info    : Done meshing 3D (0 s)
+Info    : Meshing order 2 (curvilinear on)...
+Info    : Meshing curve 1 order 2
+Info    : Meshing curve 2 order 2
+Info    : Meshing curve 3 order 2
+Info    : Meshing curve 4 order 2
+Info    : Meshing curve 5 order 2
+Info    : Meshing curve 6 order 2
+Info    : Meshing curve 7 order 2
+Info    : Meshing curve 8 order 2
+Info    : Meshing curve 9 order 2
+Info    : Meshing curve 10 order 2
+Info    : Meshing curve 11 order 2
+Info    : Meshing curve 12 order 2
+Info    : Meshing surface 14 order 2
+Info    : Meshing surface 16 order 2
+Info    : Meshing surface 18 order 2
+Info    : Meshing surface 20 order 2
+Info    : Meshing surface 22 order 2
+Info    : Meshing surface 24 order 2
+Info    : Meshing volume 26 order 2
+Info    : Finalized high order topology of periodic connections
+Info    : Done meshing order 2 (0.036 s)
+Info    : 19175 vertices 3472 elements
+Info    : Writing 'half-fuel-channel-msre-dimensions.msh'...
+Info    : Done writing 'half-fuel-channel-msre-dimensions.msh'
+Info    : Stopped on Fri Jun 30 11:55:42 2017
