+++ src/materials/required_xs.md
@@ -1,0 +1,12 @@
+The following are required cross sections for Moltres:
+
+- fissxs
+- remxs
+- nsf
+- fisse
+- diffcoef
+- recipvel
+- chi
+- gtransfxs
+- beta\_eff
+- decay\_constant
