### 2017-07-05 96deed8
added .txt extensions and FISSXS
