### 2017-07-05 975bc65
correct spelling of diffcoef
