+++ src/kernels/FissionHeatSource.C
@@ -1,3 +1,6 @@
+/* This kernel will likely only be used with k-eigenvalue calculation mode with
+ * neutrons only, since the power is normalized in this case.  */
+
 #include "FissionHeatSource.h"
 
 template <>
