+++ tutorial/step01_groupConstants/MSREProperties/msre_gentry_4gfuel_FISSE.txt
@@ -1,15 +1,0 @@
-633.0 202.746 202.27 202.27 202.27
-700.0 202.747 202.27 202.27 202.27
-800.0 202.749 202.27 202.27 202.27
-900.0 202.75 202.27 202.27 202.27
-1000.0 202.751 202.27 202.27 202.27
-1100.0 202.753 202.27 202.27 202.27
-1200.0 202.754 202.27 202.27 202.27
-1300.0 202.755 202.27 202.27 202.27
-1400.0 202.757 202.27 202.27 202.27
-1500.0 202.758 202.27 202.27 202.27
-1600.0 202.759 202.27 202.27 202.27
-1700.0 202.76 202.27 202.27 202.27
-1800.0 202.762 202.27 202.27 202.27
-1900.0 202.763 202.27 202.27 202.27
-2000.0 202.764 202.27 202.27 202.27
