+++ tutorial/step01_groupConstants/MSREProperties/msre_gentry_4gmoder_RECIPVEL.txt
@@ -1,14 +1,0 @@
-633.0 2.35193e-09 4.2325e-08 3.85617e-07 2.1832e-06
-700.0 2.35232e-09 4.2323e-08 3.86269e-07 2.13241e-06
-800.0 2.35222e-09 4.23178e-08 3.87523e-07 2.06313e-06
-900.0 2.35245e-09 4.23224e-08 3.89156e-07 2.00078e-06
-1000.0 2.3525e-09 4.23367e-08 3.91252e-07 1.9389e-06
-1100.0 2.35261e-09 4.23557e-08 3.93863e-07 1.88643e-06
-1200.0 2.35274e-09 4.23494e-08 3.97026e-07 1.83542e-06
-1300.0 2.35267e-09 4.23439e-08 4.0052e-07 1.7934e-06
-1400.0 2.35309e-09 4.2348e-08 4.04804e-07 1.75334e-06
-1500.0 2.35313e-09 4.23473e-08 4.09939e-07 1.71484e-06
-1600.0 2.35334e-09 4.23461e-08 4.16007e-07 1.67838e-06
-1700.0 2.35332e-09 4.23486e-08 4.21754e-07 1.64916e-06
-1800.0 2.35323e-09 4.23455e-08 4.28081e-07 1.62111e-06
-1900.0 2.35357e-09 4.23459e-08 4.35029e-07 1.59487e-06
