+++ tutorial/step01_groupConstants/MSREProperties/msre_gentry_4gfuel_FISSXS.txt
@@ -1,15 +1,0 @@
-633.0 0.000172839 0.00153527 0.003802 0.0223878
-700.0 0.000170353 0.00151395 0.00374785 0.0219858
-800.0 0.000166715 0.00148267 0.00367501 0.0214045
-900.0 0.000163161 0.00145325 0.00359837 0.0208385
-1000.0 0.00015967 0.00142294 0.00352555 0.0203064
-1100.0 0.000156274 0.00139442 0.00345409 0.0197876
-1200.0 0.000152935 0.00136629 0.00338531 0.0192809
-1300.0 0.000149666 0.00133727 0.00331549 0.0188015
-1400.0 0.000146478 0.00131006 0.00324824 0.0183377
-1500.0 0.000143357 0.00128374 0.0031836 0.0178845
-1600.0 0.00014029 0.00125666 0.00311816 0.0174559
-1700.0 0.000137306 0.00123126 0.00305511 0.0170365
-1800.0 0.000134388 0.00120576 0.00299362 0.0166242
-1900.0 0.000131524 0.0011806 0.00293275 0.016235
-2000.0 0.000128719 0.001156 0.00287479 0.0158518
