### 2017-07-06 bd13e2d
fission heat source comment moved to header docstring
