### 2017-07-06 ba2e937
pretty, clangy formatting
