+++ include/kernels/FissionHeatSource.h
@@ -3,14 +3,15 @@
 
 #include "Kernel.h"
 
-/** This kernel will likely only be used with k-eigenvalue calculation mode 
- * with neutrons only, since the power is normalized in this case.
- */
 class FissionHeatSource;
 
 template <>
 InputParameters validParams<FissionHeatSource>();
 
+/**
+ * This kernel will likely only be used with k-eigenvalue calculation mode
+ * with neutrons only, since the power is normalized in this case.
+ */
 class FissionHeatSource : public Kernel
 {
 public:
