### 2017-07-07 25c52d5
Reorganize analysis directories.
