+++ problems/MooseGold/property_file_dir/msr2g_part_U_full_core_fuel_data_func_of_fuel_temp_REMXS.txt
@@ -1,3 +1,0 @@
-900 6.08614E-03 2.88730E-02 
-950 6.11358E-03 2.82635E-02 
-1000 6.15380E-03 2.77188E-02 
