+++ problems/MooseGold/property_file_dir/B1_full_core_cuboid_msre_comp_fuel_data_func_of_fuel_temp_REMXS.txt
@@ -1,3 +1,0 @@
-900 4.06655E-03 2.84740E-02 
-950 4.03682E-03 2.84122E-02 
-1000 4.01034E-03 2.83784E-02 
