+++ problems/MooseGold/property_file_dir/full_core_cuboid_msre_comp_mod_data_func_of_mod_temp_DIFFCOEF.txt
@@ -1,3 +1,0 @@
-900 6.54410E-01 7.06420E-02 
-950 6.56068E-01 8.83296E-02 
-1000 6.57888E-01 1.09245E-01 
