+++ problems/MooseGold/property_file_dir/B1_full_core_cuboid_msre_comp_mod_data_func_of_mod_temp_BETA_EFF.txt
@@ -1,3 +1,0 @@
-900 2.50385E-04 1.11737E-03 6.99880E-04 1.45079E-03 2.35241E-03 6.70077E-04 6.06870E-04 1.69862E-04 
-950 2.53816E-04 1.08058E-03 7.05898E-04 1.44301E-03 2.37205E-03 6.76352E-04 6.01554E-04 1.66022E-04 
-1000 2.48330E-04 1.10759E-03 6.78547E-04 1.43422E-03 2.35394E-03 6.58824E-04 6.03894E-04 1.69026E-04 
