+++ problems/MooseGold/property_file_dir/B1_full_core_cuboid_msre_comp_mod_data_func_of_mod_temp_REMXS.txt
@@ -1,3 +1,0 @@
-900 5.40713E-02 1.32716E-03 
-950 6.21709E-02 1.65740E-03 
-1000 6.97188E-02 2.04079E-03 
