+++ problems/MooseGold/property_file_dir/B1_full_core_cuboid_msre_comp_fuel_data_func_of_fuel_temp_DIFFCOEF.txt
@@ -1,3 +1,0 @@
-900 1.30314E+00 1.24502E+00 
-950 1.31704E+00 1.25829E+00 
-1000 1.33076E+00 1.27083E+00 
