+++ problems/MooseGold/property_file_dir/full_core_cuboid_msre_comp_fuel_data_func_of_fuel_temp_DIFFCOEF.txt
@@ -1,3 +1,0 @@
-900 7.54399E-01 2.41650E+00 
-950 7.56495E-01 2.34299E+00 
-1000 7.56909E-01 2.27179E+00 
