+++ problems/MooseGold/property_file_dir/msr2g_part_U_full_core_fuel_data_func_of_fuel_temp_DIFFCOEF.txt
@@ -1,3 +1,0 @@
-900 7.97137E-01 2.28009E+00 
-950 7.99354E-01 2.21372E+00 
-1000 8.00713E-01 2.13936E+00 
