+++ problems/MooseGold/property_file_dir/B1_full_core_cuboid_msre_comp_fuel_data_func_of_fuel_temp_RECIPVEL.txt
@@ -1,3 +1,0 @@
-900 9.36273E-08 2.01973E-06 
-950 9.43496E-08 1.98448E-06 
-1000 9.51382E-08 1.95161E-06 
