+++ problems/MooseGold/property_file_dir/full_core_cuboid_msre_comp_fuel_data_func_of_fuel_temp_CHI.txt
@@ -1,3 +1,0 @@
-900 1.00000E+00 0.00000E+00 
-950 1.00000E+00 0.00000E+00 
-1000 1.00000E+00 0.00000E+00 
