+++ problems/MooseGold/property_file_dir/B1_full_core_cuboid_msre_comp_mod_data_func_of_mod_temp_DIFFCOEF.txt
@@ -1,3 +1,0 @@
-900 9.13797E-01 7.64814E-01 
-950 9.01140E-01 7.65549E-01 
-1000 8.89225E-01 7.66264E-01 
