+++ tutorial/step01_groupConstants/MSREProperties/msre_gentry_4gmoder_DIFFCOEF.txt
@@ -1,0 +1,14 @@
+633.0 1.20787 0.795616 0.793327 0.756596
+700.0 1.20915 0.796609 0.794142 0.75799
+800.0 1.21139 0.797987 0.795431 0.759372
+900.0 1.21358 0.799449 0.796814 0.760684
+1000.0 1.21568 0.800872 0.798151 0.762617
+1100.0 1.21794 0.802331 0.799489 0.763832
+1200.0 1.22009 0.803745 0.800875 0.765413
+1300.0 1.22228 0.805217 0.802219 0.76642
+1400.0 1.22434 0.806663 0.803536 0.767685
+1500.0 1.22662 0.808108 0.804837 0.769152
+1600.0 1.22874 0.809583 0.80609 0.770742
+1700.0 1.23097 0.811015 0.807425 0.771737
+1800.0 1.23323 0.812495 0.808797 0.772846
+1900.0 1.23541 0.813968 0.810083 0.774067
