+++ tutorial/step01_groupConstants/MSREProperties/msre_gentry_4gfuel_INVV.txt
@@ -1,0 +1,15 @@
+633.0 2.31881e-09 4.18652e-08 3.89292e-07 1.9966e-06
+700.0 2.31591e-09 4.18668e-08 3.89602e-07 1.98922e-06
+800.0 2.31171e-09 4.18738e-08 3.90075e-07 1.97889e-06
+900.0 2.30759e-09 4.18869e-08 3.90549e-07 1.96919e-06
+1000.0 2.30394e-09 4.19026e-08 3.9109e-07 1.96015e-06
+1100.0 2.30014e-09 4.1907e-08 3.91544e-07 1.95162e-06
+1200.0 2.29635e-09 4.19233e-08 3.92171e-07 1.94368e-06
+1300.0 2.29209e-09 4.19298e-08 3.92759e-07 1.93604e-06
+1400.0 2.28872e-09 4.19425e-08 3.93294e-07 1.92928e-06
+1500.0 2.28478e-09 4.19505e-08 3.93933e-07 1.92302e-06
+1600.0 2.28096e-09 4.19591e-08 3.94499e-07 1.9171e-06
+1700.0 2.27749e-09 4.1975e-08 3.9526e-07 1.91166e-06
+1800.0 2.27406e-09 4.19832e-08 3.95921e-07 1.90653e-06
+1900.0 2.27034e-09 4.1995e-08 3.96644e-07 1.90173e-06
+2000.0 2.26729e-09 4.1999e-08 3.97333e-07 1.8971e-06
