### 2017-07-10 b6546cd
removed old property files
