+++ tutorial/step01_groupConstants/MSREProperties/msre_gentry_4gfuel_DIFF_COEF.txt
@@ -1,15 +1,0 @@
-633.0 1.27251 1.2358 1.2361 1.15834
-700.0 1.29054 1.25321 1.25319 1.17355
-800.0 1.31843 1.27971 1.27919 1.19668
-900.0 1.34673 1.30663 1.30575 1.22039
-1000.0 1.37549 1.33432 1.33288 1.24441
-1100.0 1.40508 1.36253 1.3608 1.26902
-1200.0 1.4356 1.39131 1.38915 1.29412
-1300.0 1.46619 1.42087 1.41821 1.31963
-1400.0 1.49758 1.45109 1.44807 1.3458
-1500.0 1.52968 1.48189 1.47851 1.37238
-1600.0 1.56258 1.5133 1.5095 1.39953
-1700.0 1.59597 1.54561 1.5413 1.42727
-1800.0 1.63066 1.57837 1.57384 1.4556
-1900.0 1.66564 1.61202 1.60701 1.4843
-2000.0 1.7011 1.64646 1.64107 1.51356
