### 2017-07-10 946dafc
clang formatted roddedmaterial
