+++ src/materials/RoddedMaterial.C
@@ -1,0 +1,104 @@
+#include "RoddedMaterial.h"
+#include "GenericMoltresMaterial.h"
+#include "MooseUtils.h"
+
+template <typename GenericMoltresMaterial>
+InputParameters
+validParams<RoddedMaterial>()
+{
+  InputParameters params = validParams<RoddedMaterial>();
+  params.addRequiredParam<ScalarVariable>("rodPosition", "Position of the control rod.");
+  params.addRequiredParam<Real>("absorb_factor",
+                                "The material inherits from some other. How much more absorbing?");
+  return params;
+}
+
+RoddedMaterial::RoddedMaterial(const InputParameters & parameters)
+  : GenericMoltresMaterial(parameters),
+    _absorb_factor(getParam<Real>("absorb_factor")),
+    _rod_pos(getParam<ScalarVariable>("rodPosition"))
+{
+  if (_interp_type != "spline")
+    mooseError("Only spline works with this class. It's meant for testing.");
+}
+
+void
+RoddedMaterial::computeSplineAbsorbingQpProperties()
+{
+  for (decltype(_num_groups) i = 0; i < _num_groups; ++i)
+  {
+    _remxs[_qp][i] = _xsec_spline_interpolators["REMXS"][i].sample(_temperature[_qp]);
+    _fissxs[_qp][i] = _xsec_spline_interpolators["FISSXS"][i].sample(_temperature[_qp]);
+    _nsf[_qp][i] = _xsec_spline_interpolators["NSF"][i].sample(_temperature[_qp]);
+    _fisse[_qp][i] = _xsec_spline_interpolators["FISSE"][i].sample(_temperature[_qp]) * 1e6 *
+                     1.6e-19; // convert from MeV to Joules
+    _diffcoef[_qp][i] = _xsec_spline_interpolators["DIFFCOEF"][i].sample(_temperature[_qp]);
+    _recipvel[_qp][i] = _xsec_spline_interpolators["RECIPVEL"][i].sample(_temperature[_qp]);
+    _chi[_qp][i] = _xsec_spline_interpolators["CHI"][i].sample(_temperature[_qp]);
+    _d_remxs_d_temp[_qp][i] =
+        _xsec_spline_interpolators["REMXS"][i].sampleDerivative(_temperature[_qp]) * _absorb_factor;
+    _d_fissxs_d_temp[_qp][i] = 0.0;
+    _d_nsf_d_temp[_qp][i] =
+        _xsec_spline_interpolators["NSF"][i].sampleDerivative(_temperature[_qp]);
+    _d_fisse_d_temp[_qp][i] =
+        _xsec_spline_interpolators["FISSE"][i].sampleDerivative(_temperature[_qp]) * 1e6 *
+        1.6e-19; // convert from MeV to Joules
+    _d_diffcoef_d_temp[_qp][i] =
+        _xsec_spline_interpolators["DIFFCOEF"][i].sampleDerivative(_temperature[_qp]);
+    _d_recipvel_d_temp[_qp][i] =
+        _xsec_spline_interpolators["RECIPVEL"][i].sampleDerivative(_temperature[_qp]);
+    _d_chi_d_temp[_qp][i] =
+        _xsec_spline_interpolators["CHI"][i].sampleDerivative(_temperature[_qp]);
+  }
[…]
