### 2017-07-10 d0165a8
Add new class creation script.
