### 2017-07-10 8e852e1
Change initialization for _rod_pos to something MOOSE recognizes.
