### 2017-07-12 ca44522
Converge transient single channel blockage case.
