### 2017-07-12 cfeb1c4
Add default block restriction to PrecursorKernelAction
