### 2017-07-13 323c5ca
ok, switched back to using a scalar variable in order to *actually* do control
