### 2017-07-13 aa48600
got rod removal transient working
