### 2017-07-17 6008ba4
made diracHX, started on loss of cooling accident
