+++ problems/LOSCA/README.md
@@ -1,0 +1,4 @@
+Simulates a loss of secondary cooling accident in 2D MSRE-like system.
+ - uses multiapp to simulate primary loop as 1D problem
+ - HX is a dirac kernel that removes heat
+ - dirac HX heat removal rate is controllable, so this is set through a function to simulate LOSCA
