### 2017-07-18 3920b7e
HX failure case is "working"
