+++ src/base/MoltresApp.C
@@ -54,6 +54,7 @@
 #include "ElmIntegTotFissPostprocessor.h"
 #include "ElmIntegTotFissHeatPostprocessor.h"
 #include "ElmIntegTotFissNtsPostprocessor.h"
+#include "RealControlParameterReporter.h"
 
 // AuxKernels
 #include "FissionHeatSourceAux.h"
@@ -159,6 +160,7 @@ MoltresApp::registerObjects(Factory & factory)
   registerPostprocessor(ElmIntegTotFissHeatPostprocessor);
   registerPostprocessor(ElmIntegTotFissNtsPostprocessor);
   registerPostprocessor(AverageFissionHeat);
+  registerPostprocessor(RealControlParameterReporter);
   registerAux(FissionHeatSourceAux);
   registerAux(FissionHeatSourceTransientAux);
   registerAux(ModeratorHeatSourceTransientAux);
