### 2017-07-19 6dd5106
fixed typo in README for LOSCA
