### 2017-07-19 dcc023d
added LOSCA w walls held at inlet temp
