### 2017-07-20 1ace584
Add python script for calculating fuel pitch fraction.
