### 2017-07-20 6c31058
Update LOSCA case to run as advertised.
