+++ problems/LOSCA/README.md
@@ -5,13 +5,12 @@ Simulates a loss of secondary cooling accident in 2D MSRE-like system.
 
 # To simulate HX failure transient:
 
-1) run the input files in this directory until they come to steady-state. ctrl-c'ing out of the sim
-   may be good since the temperature appears to very gently oscillate in steady-state, thus never making
-   it to the 10000 s mark.
-2) the input in HXFailure will look in the above directory for the correct output files to kick the simulation
-   off on steady-state.
+1. Mesh the geo file: `gmsh -2 2d_lattice_structured.geo`
 
-3) The HXFailAdiabatic case is just like HXFailure, but instead of holding vessel walls at 900K, the vessel walls are
-   held at inlet temperature. This models perfect heat conduction between the downcomer and core with zero conduction
-   to the outside world. Ideally, some asymtotic temperature should be reached given that the reactor has a negative temperature
-   coefficient of reactivity for all temperatures over operating temp.
+2. Run the master input file `auto_diff_rho.i` in this directory:
+   `mpirun -np 8 moltres-opt -i auto_diff_rho.i`
+
+3. `cd` into the `HXFailure` directory and run the master input file
+   (same name: `auto_diff_rho.i`). This input file in HXFailure will look
+   in the above directory for the correct output files to kick the
+   simulation off on steady-state.
