### 2017-07-21 1632226
added 27 eV split 2 group constants with MSRE rods out
