+++ property_file_dir/serp3d2group27eV/2g27eV_fuel_REMXS.txt
@@ -1,0 +1,11 @@
+633.0 0.00468105 0.0224653
+700.0 0.00462225 0.0221649
+800.0 0.00453743 0.0217018
+900.0 0.00443937 0.0212668
+1000.0 0.00435159 0.0208291
+1100.0 0.00426954 0.0204001
+1200.0 0.00418625 0.0199787
+1300.0 0.00410355 0.0195675
+1400.0 0.0040255 0.0191648
+1500.0 0.00394193 0.0187673
+1600.0 0.00386305 0.0183844
