+++ property_file_dir/serp3d2group27eV/2g27eV_moder_INVV.txt
@@ -1,0 +1,14 @@
+633.0 2.12722e-08 1.56705e-06
+700.0 2.1271e-08 1.53984e-06
+800.0 2.12655e-08 1.50092e-06
+900.0 2.12723e-08 1.46613e-06
+1000.0 2.12813e-08 1.43058e-06
+1100.0 2.12867e-08 1.39947e-06
+1200.0 2.1282e-08 1.36867e-06
+1300.0 2.12818e-08 1.34226e-06
+1400.0 2.12854e-08 1.31702e-06
+1500.0 2.12838e-08 1.29146e-06
+1600.0 2.12832e-08 1.26679e-06
+1700.0 2.12837e-08 1.24617e-06
+1800.0 2.12836e-08 1.22593e-06
+1900.0 2.12813e-08 1.20602e-06
