+++ property_file_dir/serp3d2group27eV/2g27eV_moder_CHIT.txt
@@ -1,0 +1,14 @@
+633.0 0.0 0.0
+700.0 0.0 0.0
+800.0 0.0 0.0
+900.0 0.0 0.0
+1000.0 0.0 0.0
+1100.0 0.0 0.0
+1200.0 0.0 0.0
+1300.0 0.0 0.0
+1400.0 0.0 0.0
+1500.0 0.0 0.0
+1600.0 0.0 0.0
+1700.0 0.0 0.0
+1800.0 0.0 0.0
+1900.0 0.0 0.0
