+++ property_file_dir/serp3d2group27eV/2g27eV_fuel_DIFFCOEF.txt
@@ -1,0 +1,11 @@
+633.0 1.25565 1.18562
+700.0 1.27364 1.20132
+800.0 1.30068 1.22511
+900.0 1.32833 1.24948
+1000.0 1.35686 1.27424
+1100.0 1.38577 1.29959
+1200.0 1.41525 1.32551
+1300.0 1.44559 1.35195
+1400.0 1.47645 1.37875
+1500.0 1.508 1.40646
+1600.0 1.54044 1.43434
