### 2017-07-24 0ac4aef
GCs for 3D MSRE in serp; 2G split at 0.78eV
