+++ property_file_dir/serp3d2group0.78eV/2g0.78eV_fuel_REMXS.txt
@@ -1,0 +1,11 @@
+633.0 0.00604105 0.0283882
+700.0 0.00599819 0.0279683
+800.0 0.00592779 0.0273708
+900.0 0.00585463 0.0268073
+1000.0 0.00578651 0.0262755
+1100.0 0.00571381 0.0257604
+1200.0 0.0056404 0.0252772
+1300.0 0.00556414 0.024814
+1400.0 0.00548958 0.0243536
+1500.0 0.00542436 0.0239225
+1600.0 0.0053475 0.0235287
