+++ property_file_dir/serp3d2group0.78eV/2g0.78eV_fuel_SP0.txt
@@ -1,0 +1,11 @@
+633.0 0.296397 0.00194731 0.000648272 0.273251
+700.0 0.292289 0.00193033 0.00073498 0.269777
+800.0 0.286271 0.00190711 0.000861067 0.264642
+900.0 0.28036 0.00188297 0.000989195 0.259567
+1000.0 0.274584 0.00186308 0.00111284 0.25461
+1100.0 0.268914 0.00184369 0.00123675 0.249704
+1200.0 0.26335 0.00182555 0.00136358 0.244861
+1300.0 0.257901 0.00180627 0.00148872 0.240131
+1400.0 0.252581 0.00178933 0.00161204 0.23548
+1500.0 0.247331 0.00177429 0.00173796 0.230885
+1600.0 0.242218 0.00175645 0.00186373 0.226385
