+++ property_file_dir/serp3d2group0.78eV/2g0.78eV_fuel_FISS.txt
@@ -1,0 +1,11 @@
+633.0 0.00133577 0.0222019
+700.0 0.00131644 0.0218047
+800.0 0.00128957 0.0212251
+900.0 0.00126248 0.0206632
+1000.0 0.00123634 0.020142
+1100.0 0.00121081 0.0196251
+1200.0 0.0011863 0.0191241
+1300.0 0.00116141 0.0186522
+1400.0 0.00113752 0.018191
+1500.0 0.00111482 0.0177426
+1600.0 0.00109178 0.0173193
