+++ property_file_dir/serp3d2group0.78eV/2g0.78eV_moder_REMXS.txt
@@ -1,0 +1,14 @@
+633.0 0.00358073 0.000999721
+700.0 0.00362995 0.00114802
+800.0 0.00371508 0.00139398
+900.0 0.00381609 0.00167629
+1000.0 0.00394807 0.00204049
+1100.0 0.00410533 0.00246126
+1200.0 0.00431395 0.00300825
+1300.0 0.00453944 0.00360128
+1400.0 0.0048214 0.00433325
+1500.0 0.005171 0.00522648
+1600.0 0.00558966 0.00629533
+1700.0 0.00599551 0.00734296
+1800.0 0.00645041 0.00852619
+1900.0 0.00696703 0.00986387
