+++ property_file_dir/serp3d2group0.78eV/2g0.78eV_fuel_DIFFCOEF.txt
@@ -1,0 +1,11 @@
+633.0 1.25194 1.1594
+700.0 1.26963 1.17461
+800.0 1.2965 1.19784
+900.0 1.32401 1.22164
+1000.0 1.35209 1.24572
+1100.0 1.38086 1.27036
+1200.0 1.41032 1.29556
+1300.0 1.44041 1.32108
+1400.0 1.47101 1.34729
+1500.0 1.50246 1.37395
+1600.0 1.53451 1.40114
