### 2017-07-25 67a6f69
changed PrecursorKernel action to Precursors in all files
