### 2017-07-26 6ea1fbc
Move PrecursorKernelAction to PrecursorAction
