### 2017-07-27 82a2a35
initial work on natural convection is done
