### 2017-07-27 05a64eb
Add some visualization with yt examples
