### 2017-07-28 ed8ebcf
boussinesq now looks reasonable on the 2D square example
