### 2017-07-28 c070d80
made Alex's recommended changes from PR
