### 2017-07-28 970bd8a
Add python scripts for generating MSRE comparison figures.
