### 2017-08-02 #298 @katyhuff
@scopatz , just wanted to check in about whether the docker image has worked for you -- it should get you to running the tests in no time. Thanks so much for being willing to review!
