### 2017-08-02 c81f48c
Update python files to show how to add external data to yt generated plot.
