+++ problems/constant_inlet_outlet_temp/fuel_vs_mod_temp.py
@@ -8,7 +8,8 @@ def _units(field, data):
 
 
 ds = yt.load(
-    '/home/lindsayad/projects/moltres/problems/constant_inlet_outlet_temp/2group.e', step=-1)
+    '/home/lindsayad/projects/moltres/problems/'
+    'constant_inlet_outlet_temp/2group.e', step=-1)
 ds.add_field(('all', 'temp_K'), units='K', function=_units, take_log=False)
 lines = []
 lines.append(yt.LineBuffer(
@@ -23,10 +24,11 @@ plt.set_x_unit('inch')
 plt.set_unit(('all', 'temp_K'), 'degF')
 plt.set_xlabel('z (in)')
 plt.set_ylabel('Temperature (F)')
-plt.show()
+plt._setup_plots()
 
 msre_temp_axial_data = np.loadtxt(
-    '/home/lindsayad/publications/figures/msre_axial_temps.csv', skiprows=1, delimiter=',')
+    '/home/lindsayad/publications/figures/'
+    'msre_axial_temps.csv', skiprows=1, delimiter=',')
 x = msre_temp_axial_data[:, 0]
 fuel_temp = msre_temp_axial_data[:, 1]
 graphite_temp = msre_temp_axial_data[:, 2]
@@ -34,6 +36,7 @@ plt.plots[('all', 'temp_K')].axes.plot(x, fuel_temp, label=r'fuel msre')
 plt.plots[('all', 'temp_K')].axes.plot(
     x, graphite_temp, label=r'graphite msre')
 plt.plots[('all', 'temp_K')].axes.legend()
-plt.show()
+plt._setup_plots()
 
-plt.save('/home/lindsayad/publications/figures/combined_msre_moltres_axial_temps.eps')
+plt.save('/home/lindsayad/publications/figures/'
+         'combined_msre_moltres_axial_temps.eps')
