+++ include/auxkernels/MatDiffusionAux.h
@@ -10,6 +10,14 @@ class MatDiffusionAux;
 template <>
 InputParameters validParams<MatDiffusionAux>();
 
+/**
+ * Computes the Euclidean norm of the flux of some diffusing variable at
+ * any point in space. In other words, this computes:
+ * \f[
+ *   | D \nabla f |
+ * \f]
+ * Where D is the diffusion coefficient and f is some scalar variable in the simulation.
+ */
 class MatDiffusionAux : public AuxKernel
 {
 public:
