### 2017-08-08 d1ca109
moved docstring to above class declaration
