### 2017-08-08 68dced6
made PR recommended changes
