+++ include/kernels/CoupledFissionKernel.h
@@ -12,7 +12,7 @@ InputParameters validParams<CoupledFissionKernel>();
 
 /**
  * Computes fission source of neutrons without normalizing by
- * \f$ 1/k \f$.
+ * \f$ 1/k \f$. Note that this kernel is meant for transients.
  */
 class CoupledFissionKernel : public Kernel, public ScalarTransportBase
 {
