### 2017-08-09 e0f30cd
finished finding multiplicative value to match ORNL rod worth
