+++ problems/rodMatchSerpentWorth/README.md
@@ -1,0 +1,8 @@
+The solve here is only done in eigenvalue mode. The objective is to change the absorbtion of the control rod until
+the same reactivity worth experienced at the MSRE is obtained in Moltres.
+
+In this top directory, the radius of the 2D case found the match Serpent's unrodded reactivity was found to be 78.58cm.
+Serpent found k=1.045 +- 0.0010. This simulation yields k= 1.045177e+00.
+
+In the adjustAbsorb directory, the control rod's absorbtion factor was found to match the MSRE results. To reproduce results,
+the .msh file generated by gmsh must be copied into that directory.
