### 2017-08-28 fbc9ba7
prec action now add postproc, bcs, transf for circulation
