### 2017-08-29 854c1ca
removed inheriting _multi_app in PrecursorAction
