### 2017-08-29 #298 @scopatz
Hi @lindsayad - I am now seeing this:

```
scopatz@artemis ~ $ docker pull lindsayad/moltres
Using default tag: latest
Error response from daemon: repository lindsayad/moltres not found: does not exist or no pull access
```

Please advise.
