### 2017-08-30 b2d0e25
add task dependency to MoltresApp
