### 2017-08-30 3df0910
circulating precursors now works
