+++ include/actions/PrecursorAction.h
@@ -35,9 +35,6 @@ protected:
 
   /// optional object name suffix
   std::string _object_suffix;
-
-  /// name of multi app for precursor circulation
-  MultiAppName _multi_app;
 };
 
 template <>
