### 2017-08-30 0cb971e
PR comments considered
