### 2017-08-30 #298 @lindsayad
I gave you bad information. Instead do:
```bash
docker pull lindsayad/moose
```
and then once you're in an interactive terminal:
```
su - moose-user
```
