### 2017-08-30 #298 @lindsayad
You should then be able to
```
cd ~/projects/moltres
./run_tests -j8
```
