### 2017-09-01 #298 @scopatz
> References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?

The Moose reference does not list a DOI
