### 2017-09-01 #298 @scopatz
> Have the functional claims of the software been confirmed?

How do I actually run the example? This is not presented anywhere in the documentation.

> Have any performance claims of the software been confirmed?

likewise can you please give an example of how to run with parallelism
