### 2017-09-01 #298 @scopatz
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Similarly, there is no full walkthrough of inputs to results that I could find. The [example input](http://arfc.github.io/software/moltres/wiki/input_example/) page isn't enough for me to run the problem and there are no results to verify against
