### 2017-09-01 #298 @scopatz
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g. API method documentation)?

Most of the API lacks any form of in-line documentation that is translated to doxygen.

Also, the API docs are not clearly labeled as such on the website. Instead, it says that the, "Doxygen pages are here." This requires that the user have a priori knowledge that doxygen is a tool for producing documentation
