### 2017-09-01 #298 @scopatz
Hi @lindsayad, sorry it took me so long to get to the review!  Please let me know if/when you have a chance to get to the issues I have brought up above. Thanks!
