### 2017-09-05 5497b6c
Change kernel_name to bc_name where appropriate and clang-format.
