### 2017-09-05 2f25b02
Split main act function into pieces.
