+++ src/actions/PrecursorAction.C
@@ -88,9 +88,7 @@ PrecursorAction::act()
 
     if (getParam<bool>("create_vars"))
     {
-      //
       // See whether we want to use an old solution
-      //
       if (getParam<bool>("init_from_file"))
       {
         if (_current_task == "check_copy_nodal_vars")
@@ -103,251 +101,269 @@ PrecursorAction::act()
         }
       }
 
-      //
       // Create variable names
-      //
-
       if (_current_task == "add_variable")
         addVariable(var_name);
     }
 
+    // kernels
     if (_current_task == "add_kernel")
-    {
-      // Set up PrecursorSource kernels
+      kernelAct(op, var_name);
 
-      {
-        InputParameters params = _factory.getValidParams("PrecursorSource");
-        params.set<NonlinearVariableName>("variable") = var_name;
-        params.set<unsigned int>("num_groups") = _num_groups;
-        params.set<unsigned int>("precursor_group_number") = op;
-        std::vector<std::string> include = {"temperature", "group_fluxes"};
-        params.applySpecificParameters(parameters(), include);
-        if (isParamValid("kernel_block"))
-          params.set<std::vector<SubdomainName>>("block") =
-              getParam<std::vector<SubdomainName>>("kernel_block");
-        else if (isParamValid("block"))
-          params.set<std::vector<SubdomainName>>("block") =
-              getParam<std::vector<SubdomainName>>("block");
-        params.set<bool>("use_exp_form") = getParam<bool>("nt_exp_form");
-
-        std::string kernel_name = "PrecursorSource_" + var_name + "_" + _object_suffix;
-        _problem->addKernel("PrecursorSource", kernel_name, params);
-      }
-
-      //
-      // Set up PrecursorDecay kernels
-      //
-
-      {
-        InputParameters params = _factory.getValidParams("PrecursorDecay");
-        params.set<NonlinearVariableName>("variable") = var_name;
-        params.set<unsigned int>("precursor_group_number") = op;
-        std::vector<std::string> include = {"temperature"};
-        params.applySpecificParameters(parameters(), include);
-        if (isParamValid("kernel_block"))
-          params.set<std::vector<SubdomainName>>("block") =
-              getParam<std::vector<SubdomainName>>("kernel_block");
-        else if (isParamValid("block"))
-          params.set<std::vector<SubdomainName>>("block") =
-              getParam<std::vector<SubdomainName>>("block");
-        params.set<bool>("use_exp_form") = false;
-
-        std::string kernel_name = "PrecursorDecay_" + var_name + "_" + _object_suffix;
-        _problem->addKernel("PrecursorDecay", kernel_name, params);
-      }
-
-      //
-      // Set up TimeDerivative kernels
-      //
-      if (getParam<bool>("transient"))
-      {
-        InputParameters params = _factory.getValidParams("ScalarTransportTimeDerivative");
-        params.set<NonlinearVariableName>("variable") = var_name;
-        params.set<bool>("implicit") = true;
-        if (isParamValid("kernel_block"))
-          params.set<std::vector<SubdomainName>>("block") =
-              getParam<std::vector<SubdomainName>>("kernel_block");
-        else if (isParamValid("block"))
-          params.set<std::vector<SubdomainName>>("block") =
-              getParam<std::vector<SubdomainName>>("block");
-        params.set<bool>("use_exp_form") = false;
-
-        std::string kernel_name =
-            "ScalarTransportTimeDerivative_" + var_name + "_" + _object_suffix;
[…]
