### 2017-09-06 8c29473
Remove intlet and outlet averages from outputs.
