### 2017-09-07 #298 @kyleniemeyer
To avoid any potential (or the appearance of) conflict of interest, I'd like to find a second reviewer for this submission—not necessarily someone in the domain, but mainly someone quite familiar with C++. @jochym @dvalters @pragyansmita @hugoledoux would any of you be interested in providing a second review?

(@scopatz you are doing fine, keep doing your thing 😄)
