### 2017-09-07 #298 @dvalters
@kyleniemeyer I can give it a second review if domain-specific knowledge is not required.
