### 2017-09-07 #298 @kyleniemeyer
@dvalters excellent, thanks! Even though @scopatz checked off items in the checklist, please check whether all items are satisfied from your perspective.
