### 2017-09-11 #298 @lindsayad
@scopatz I've updated our [home wiki page](http://arfc.github.io/software/moltres/) so that API docs are clearly labeled as such, instead of "Doxygen page". I've also added a significant amount of content to our [example](http://arfc.github.io/software/moltres/wiki/input_example/) which describes how to actually run the example, where you can compare to some example output, and what some variables are that you can tune and what you might expect their effect to be. I'll be working on our API documentation next.
