### 2017-09-11 #298 @lindsayad
Oops, realized I haven't updated the docker container yet. I'll do that and then report back
