### 2017-10-01 #298 @lindsayad
Docker container finally updated. Access through:
```
docker pull lindsayad/moose:0.2.0
```
