+++ problems/publication_level_cases/manuBench/groupconstants/NSF.txt
@@ -1,0 +1,1 @@
+900 0.31780612 0.27663047 0.37039145 0.6277296 1.30504206 3.52626536
