+++ problems/publication_level_cases/manuBench/groupconstants/GTRANSFXS.txt
@@ -1,0 +1,2 @@
+900 10.8476 0 0 0 0 0 5.23316 18.3666 0 0 0 0 0.401805 3.19138 29.8293 0 0 0 0.0109869 0.00234218 1.6347 21.7472 0 0 0.0025329 0.000225259 0.00170575 1.90243 22.7173 0 0.000378334 2.00405e-05 0.000124625 1.36858e-06 1.05885 23.7826
+ 
