+++ problems/publication_level_cases/manuBench/groupconstants/DECAY_CONSTANT.txt
@@ -1,0 +1,1 @@
+900 0.0124667 0.0282917 0.0425244 0.133042 0.292467 0.666488 1.63478 3.5546
