+++ problems/publication_level_cases/manuBench/groupconstants/BETA_EFF.txt
@@ -1,0 +1,1 @@
+900 0.000185153514 0.000837669377 0.000560127355 0.00120542443 0.00199766231 0.000622770723 0.000597618812 0.0001642607 0.00617068722
