### 2017-10-27 5a90838
compiles with chi_d functionality
