+++ src/materials/GenericMoltresMaterial.C
@@ -54,6 +54,7 @@ GenericMoltresMaterial::GenericMoltresMaterial(const InputParameters & parameter
     _diffcoef(declareProperty<std::vector<Real>>("diffcoef")),
     _recipvel(declareProperty<std::vector<Real>>("recipvel")),
     _chi(declareProperty<std::vector<Real>>("chi")),
+    _chi_d(declareProperty<std::vector<Real>>("chi_d")),
     _gtransfxs(declareProperty<std::vector<Real>>("gtransfxs")),
     _beta_eff(declareProperty<std::vector<Real>>("beta_eff")),
     _beta(declareProperty<Real>("beta")),
@@ -66,6 +67,7 @@ GenericMoltresMaterial::GenericMoltresMaterial(const InputParameters & parameter
     _d_diffcoef_d_temp(declareProperty<std::vector<Real>>("d_diffcoef_d_temp")),
     _d_recipvel_d_temp(declareProperty<std::vector<Real>>("d_recipvel_d_temp")),
     _d_chi_d_temp(declareProperty<std::vector<Real>>("d_chi_d_temp")),
+    _d_chi_d_d_temp(declareProperty<std::vector<Real>>("d_chi_d_temp")),
     _d_gtransfxs_d_temp(declareProperty<std::vector<Real>>("d_gtransfxs_d_temp")),
     _d_beta_eff_d_temp(declareProperty<std::vector<Real>>("d_beta_eff_d_temp")),
     _d_beta_d_temp(declareProperty<Real>("d_beta_d_temp")),
@@ -91,6 +93,7 @@ GenericMoltresMaterial::GenericMoltresMaterial(const InputParameters & parameter
                                       "DIFFCOEF",
                                       "RECIPVEL",
                                       "CHI",
+                                      "CHI_D",
                                       "GTRANSFXS",
                                       "BETA_EFF",
                                       "DECAY_CONSTANT"};
@@ -117,6 +120,7 @@ GenericMoltresMaterial::GenericMoltresMaterial(const InputParameters & parameter
     _file_map["FISSE"] = "KAPPA";
     _file_map["RECIPVEL"] = "INVV";
     _file_map["CHI"] = "CHIT";
+    _file_map["CHI_D"] = "CHID";
     _file_map["GTRANSFXS"] = "SP0";
     _file_map["DECAY_CONSTANT"] = "LAMBDA";
   }
@@ -127,6 +131,7 @@ GenericMoltresMaterial::GenericMoltresMaterial(const InputParameters & parameter
     _file_map["FISSE"] = "FISSE";
     _file_map["RECIPVEL"] = "RECIPVEL";
     _file_map["CHI"] = "CHI";
+    _file_map["CHI_D"] = "CHI_D";
     _file_map["GTRANSFXS"] = "GTRANSFXS";
     _file_map["DECAY_CONSTANT"] = "DECAY_CONSTANT";
   }
@@ -418,6 +423,7 @@ GenericMoltresMaterial::leastSquaresConstruct(std::string & property_tables_root
   _diffcoeff_consts = xsec_map["DIFFCOEF"];
   _recipvel_consts = xsec_map["RECIPVEL"];
   _chi_consts = xsec_map["CHI"];
+  _chi_d_consts = xsec_map["CHI_D"];
   _gtransfxs_consts = xsec_map["GTRANSFXS"];
[…]
