+++ problems/publication_level_cases/manuBench/groupconstants/REMXS.txt
@@ -1,0 +1,2 @@
+16.5512 21.7253 31.8009 24.2093 25.0351 27.2159
+
