### 2017-11-04 639e8bb
manubox runs but converges to trivial solution
