+++ src/kernels/ManuHX.C
@@ -1,0 +1,33 @@
+#include "ManuHX.h"
+
+template <>
+InputParameters
+validParams<ManuHX>()
+{
+  InputParameters params = validParams<Kernel>();
+  params.addClassDescription("A uniformly applied heat removal proportional to a difference"
+                             "from a nominal temperature. A global heat exchanger.");
+  params.addRequiredParam<Real>("htc", "heat transfer coefficient for q'=h*(T-T_ref)");
+  params.addRequiredParam<Real>("tref", "reference temperature");
+  return params;
+}
+
+ManuHX::ManuHX(const InputParameters & parameters)
+  : Kernel(parameters),
+    _htc(getParam<Real>("htc")),
+    _tref(getParam<Real>("tref"))
+{
+}
+
+Real
+ManuHX::computeQpResidual()
+{
+  // positive -> heat removal
+  return _test[_i][_qp] * _htc * (_u[_qp] - _tref);
+}
+
+Real
+ManuHX::computeQpJacobian()
+{
+  return _test[_i][_qp] * _htc;
+}
