### 2017-11-06 8d3f1ff
now also normalizes lamba*C source by 1/k; still no luck
