+++ problems/publication_level_cases/manuBench/note.md
@@ -1,0 +1,1 @@
+Appears that keff is far lower than it should be, around 0.02. Going to be looking into this.
