### 2017-11-07 55db92f
double checked for transposed SP0 matrix
