### 2017-11-09 7f56c3a
made nts only case. found REMXS was defined slightly differently but fixed it. not working
