+++ problems/publication_level_cases/manuBench/groupconstants/REMXS.txt
@@ -1,2 +1,1 @@
-16.5512 21.7253 31.8009 24.2093 25.0351 27.2159
-
+900 5.7036  3.3587  1.9716  2.4621  2.3178  3.4333
