### 2017-11-09 1bb0dfb
removed beta_tot so that beta_eff data is consistent
