### 2017-11-23 16a8c20
chained else if -> switch case
