### 2017-11-24 e886555
chi_d functionality now backwards compatible with tests
