### 2017-11-24 13087cb
switch case still giving trouble, see slack
