### 2017-11-24 f891c4d
add breaks to switch-case blocks
