### 2017-11-27 #298 @kyleniemeyer
Hi @scopatz, did you have a chance to complete/continue your review?
