### 2017-12-07 #298 @scopatz
Alright, sorry for the delay. I think that this is good to go, from my perspective.  Thanks for fixing this up @lindsayad!
