### 2017-12-11 #298 @dvalters
I am nearly finished with the 2nd review - apologies for delay, will complete today - don't foresee any problems though, agree with @scopatz's review.
