### 2017-12-18 #298 @dvalters
@kyleniemeyer Sorry - all done now! I am happy it meets all the requirements and agree with @scopatz's review. A well put together piece of software 🙂
