### 2017-12-18 #298 @kyleniemeyer
Hi @lindsayad @katyhuff, my only remaining comment is that the article is a bit short, and is missing some of the aspects we normally look for (http://joss.theoj.org/about#author_guidelines)

It's only about 100 words right now (we expect more like 250–1000), and the only citations are to the software itself (which isn't really necessary/appropriate, since the article and software are already associated) and then the MOOSE software. However, are there MOOSE-related publications that could be cited that describe the theory behind it? Are there articles that have used Moltres that you can cite, or those that describe the theory better?

The article could also better describe example uses/applications.

There are a number of good examples, but one I recommend you take a look at is PyGBe-LSPR (http://joss.theoj.org/papers/10.21105/joss.00306).
