### 2018-01-02 01597b2
Migrate to smart pointers
