### 2018-01-02 6d64354
Migrate to smart pointers
