### 2018-01-03 f3281b8
Update paper.
