+++ paper/paper.bib
@@ -1,18 +1,54 @@
-@online{moose,
-  author = {Gaston, Derek R. and Permann, Cody J. and Peterson, John W. and
-                  Slaughter, Andrew E. and Andrš, David and Wang, Yaqi and
-                  Short, Michael P. and Perez, Danielle M. and Tonks, Michael
-                  R. and Ortensi, Javier and Zou, Ling and Martineau, Richard
-                  C.},
-  title = {MOOSE: Multiphysics Object Oriented Simulation
-                  Environment},
-  url = {https://github.com/idaholab/moose}
+@article{moose,
+	series = {Multi-{Physics} {Modelling} of {LWR} {Static} and {Transient} {Behaviour}},
+	title = {Physics-based multiscale coupling for full core nuclear reactor simulation},
+	volume = {84},
+	issn = {0306-4549},
+	url = {http://www.sciencedirect.com/science/article/pii/S030645491400543X},
+	doi = {10.1016/j.anucene.2014.09.060},
+	abstract = {Numerical simulation of nuclear reactors is a key technology in the quest for improvements in efficiency, safety, and reliability of both existing and future reactor designs. Historically, simulation of an entire reactor was accomplished by linking together multiple existing codes that each simulated a subset of the relevant multiphysics phenomena. Recent advances in the MOOSE (Multiphysics Object Oriented Simulation Environment) framework have enabled a new approach: multiple domain-specific applications, all built on the same software framework, are efficiently linked to create a cohesive application. This is accomplished with a flexible coupling capability that allows for a variety of different data exchanges to occur simultaneously on high performance parallel computational hardware. Examples based on the KAIST-3A benchmark core, as well as a simplified Westinghouse AP-1000 configuration, demonstrate the power of this new framework for tackling{\textemdash}in a coupled, multiscale manner{\textemdash}crucial reactor phenomena such as CRUD-induced power shift and fuel shuffle.},
+	urldate = {2017-04-10},
+	journal = {Annals of Nuclear Energy},
+	author = {Gaston, Derek R. and Permann, Cody J. and Peterson, John W. and Slaughter, Andrew E. and Andr{\v s}, David and Wang, Yaqi and Short, Michael P. and Perez, Danielle M. and Tonks, Michael R. and Ortensi, Javier and Zou, Ling and Martineau, Richard C.},
+	month = oct,
+	year = {2015},
+	keywords = {Multiphysics, Full core reactor simulation, Multiphysics coupling},
+	pages = {45--54},
+	file = {ScienceDirect Full Text PDF:/home/huff/Zotero/storage/SGQXWGJV/Gaston et al. - 2015 - Physics-based multiscale coupling for full core nu.pdf:application/pdf;ScienceDirect Snapshot:/home/huff/Zotero/storage/PI8FZ93W/S030645491400543X.html:text/html}
 }
 
 @online{moltres,
-  author = {Alexander Lindsay and Kathryn Huff},
-  title = {Moltres: Open Source Simulation of Molten Salt Reactors},
-  url = {https://github.com/arfc/moltres},
-  doi = {10.5281/zenodo.801822},
-  year = 2016
+  author = {Lindsay, Alexander and Ridley, Gavin, and Rykhlevskii, Andrei, and
+                  Huff, Kathryn},
+  title = {Introduction to Moltres: an Application for Simulation of Molten Salt
+                  Reactors},
+  journaltitle = {Annals of Nuclear Energy},
+  note = {Accepted for publication},
+  year = {2017}
+}
+
+ @TechReport{petsc,
+            author = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Jed Brown and Peter Brune
+                      and Kris Buschelman and Lisandro Dalcin and Victor Eijkhout and William~D. Gropp
[…]
