### 2018-01-03 #298 @lindsayad
@kyleniemeyer I've fleshed out our article a bit more, so hoping it's ready to go!
