### 2018-01-03 #298 @kyleniemeyer
Hi @lindsayad, it looks like the `kophazi_development_2009` reference is missing the DOI, and the `moltres` reference also seems incomplete. (If the latter is accepted, perhaps update the year to 2018?)
