### 2018-01-05 #298 @lindsayad
What do you mean by incomplete? I can't provide a volume or issue number since it hasn't been published yet. I've updated the year (since it will be published in 2018 -> good catch!). Should I just not have that reference?
