### 2018-01-05 #298 @lindsayad
Kophazi reference has been updated with DOI
