### 2018-01-05 #298 @kyleniemeyer
Hi @lindsayad, ok, got it, I didn't realize at first that the `moltres` reference was something submitted.

If it hasn't yet been accepted in the journal, it's probably not appropriate to include the journal name. Instead, perhaps replace the journal name with "in review"? Also, is there a preprint link you could share? If so, it might be good to change the reference type to `misc` and add that link in the `howpublished` field.
