### 2018-01-06 #298 @lindsayad
It has been accepted. I just finished reviewing the proofs. But not published yet. Hmm good thought on the preprint. We didn't submit to arxiv but probably should anyways. Would you prefer that to the accepted but not yet published reference?

> On Jan 5, 2018, at 3:47 PM, Kyle Niemeyer <notifications@github.com> wrote:
>
> Hi @lindsayad, ok, got it, I didn't realize at first that the moltres reference was something submitted.
>
> If it hasn't yet been accepted in the journal, it's probably not appropriate to include the journal name. Instead, perhaps replace the journal name with "in review"? Also, is there a preprint link you could share? If so, it might be good to change the reference type to misc and add that link in the howpublished field.
>
