### 2018-01-06 #298 @lindsayad
I'm ready to be done but I'll let @katyhuff sign off first

> On Jan 6, 2018, at 2:23 PM, Kyle Niemeyer <notifications@github.com> wrote:
>
> @lindsayad Ah, ok. Since it's been accepted, then it's fine to include the journal name. It's up to you if you want to submit to arXiv, but if you do, then you could add the link (once it's available).
>
> I'm fine to accept this now, but it's up to you (and @katyhuff) if you want to add an arXiv link, or even wait for the DOI to become available.
>
