### 2018-01-08 #298 @lindsayad
Yes, the DOI should be 10.5281/zenodo.887941

On Sun, Jan 7, 2018 at 7:53 PM, Arfon Smith <notifications@github.com>
wrote:

> @lindsayad <https://github.com/lindsayad> - At this point could you make
> an archive of the reviewed software in Zenodo/figshare/other service and
> update this thread with the DOI of the archive? I can then move forward
> with accepting the submission.
>
