### 2018-01-08 0a281aa
Update moltres citation to published version.
