+++ paper/paper.bib
@@ -16,14 +16,24 @@
 	file = {ScienceDirect Full Text PDF:/home/huff/Zotero/storage/SGQXWGJV/Gaston et al. - 2015 - Physics-based multiscale coupling for full core nu.pdf:application/pdf;ScienceDirect Snapshot:/home/huff/Zotero/storage/PI8FZ93W/S030645491400543X.html:text/html}
 }
 
-@online{moltres,
-  author = {Lindsay, Alexander and Ridley, Gavin, and Rykhlevskii, Andrei, and
-                  Huff, Kathryn},
-  title = {Introduction to Moltres: an Application for Simulation of Molten Salt
-                  Reactors},
-  journaltitle = {Annals of Nuclear Energy},
-  note = {Accepted for publication},
-  year = {2018}
+@article{moltres,
+title = "Introduction to Moltres: An application for simulation of Molten Salt Reactors ",
+journal = "Annals of Nuclear Energy ",
+volume = "114",
+number = "",
+pages = "530 - 540",
+year = "2018",
+note = "",
+issn = "0306-4549",
+doi = "https://doi.org/10.1016/j.anucene.2017.12.025",
+url = "https://www.sciencedirect.com/science/article/pii/S0306454917304760",
+author = "Alexander Lindsay and Gavin Ridley and Andrei Rykhlevskii and Kathryn Huff",
+keywords = "Molten Salt Reactors",
+keywords = "Multiphysics",
+keywords = "Parallel computing",
+keywords = "Finite elements",
+keywords = "Reactor physics",
+keywords = "MOOSE "
 }
 
  @TechReport{petsc,
