### 2018-01-08 #298 @lindsayad
Ok, our article was actually published this morning, so if it's not too
late, I've updated our bibliography with its DOI and publication
information.

On Mon, Jan 8, 2018 at 7:10 AM, Alexander Lindsay <alexlindsay239@gmail.com>
wrote:

> Yes, the DOI should be 10.5281/zenodo.887941
>
> On Sun, Jan 7, 2018 at 7:53 PM, Arfon Smith <notifications@github.com>
> wrote:
>
>> @lindsayad <https://github.com/lindsayad> - At this point could you make
>> an archive of the reviewed software in Zenodo/figshare/other service and
>> update this thread with the DOI of the archive? I can then move forward
>> with accepting the submission.
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/298#issuecomment-355875901>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/AJxgcE6dPYerDRgPQ4ypeU-398f2OkApks5tIYMpgaJpZM4N87MR>
>> .
>>
>
>
