### 2018-01-08 #298 @lindsayad
Updated DOI for the moltres repo: 10.5281/zenodo.1137726
