### 2018-01-08 #298 @arfon
@scopatz - many thanks for your review here and to @kyleniemeyer for editing this one ✨
