### 2018-01-08 #298 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
