### 2018-01-08 #298 @kyleniemeyer
hey @lindsayad @arfon looks like the DOI in that new reference got messed up—it looks like "https://doi.org/https://doi.org/10.1016/j.anucene.2017.12.025"
