### 2018-01-08 #298 @arfon
@kyleniemeyer @lindsayad - this PR fixes the issue. The paper template expects 'bare' DOIs, i.e. without the full URL: https://github.com/arfc/moltres/pull/61
