### 2018-01-08 #298 @arfon
I've updated the paper based on https://github.com/arfc/moltres/pull/61. It might take a while before it looks fixed on the JOSS site as the papers are served from a content distribution network (CDN).
